modify retriever_eval nb to use cohere re rank metric
